+++ Server/system.Business/Services/AttachmentService.cs
@@ -14,7 +14,7 @@ public class AttachmentService(
     IWorkerRepository workerRepository,
     ILogger<AttachmentService> logger) : IAttachmentService
 {
-    public async Task<AttachmentItem> GetAttachmentById(int id, CancellationToken ct)
+    public async Task<AttachmentItem> GetAttachmentById(long id, CancellationToken ct)
     {
         return await attachmentRepository.GetById(id, ct)
                ?? throw new NotFoundException($"Attachment {id} not found");
